how to save gene exoression matrix as .csv file
I made it in a way so that  if the .csv file is opened, the actual table with the right columns can be seen.
I had some problems reading the .csv file because of the row and column names. If the former style of the matrix has to be restored please follow the code I wrote there.